fix(sidebar): incorrect tabs initialization when multiple groups
When tabs are interleaved in multiple groups (e.g. one in the first
group, then one in the second group, then two in the first group again)
only the last group of tabs from a group (in our example, the last two
for the first group) would be properly displayed in the sidebar, causing
more issue down the line (e.g. opening a new tab wouldn't be shown).

This was caused by `groupBy` only grouping consecutive instance of equal
values. This was fixed by sorting the array before.