fix: tab activation happening before creation due to groups restoration logic
The previous iteration to correctly restore tabs was on tab creation to
launch a fiber that will read the tabValue and windowValue, and only
then send the message to the sidebar. This was necessary because
getTabValue/getWindowValue are both async.
The issue was that the activation message would then arrive *before* the
tab creation message; indeed, the activation is done in Effect while the
other isn't, delaying the tab creation.

The solution is to delay only the groups initialization and assignation.
A tab is first added to the sidebar, defaulting to the current group. A
new message is then sent to assign the tab to its correct group.