refactor(Sidebar): remove Purescript hover logic, rely on :hover instead
The workaround (implementing :hover in Purescript) was necessary because
during drag-and-drop the hover effect would affect the tab at the
position of the selected tab.

The solution is to add a `is-not-dragging` class to the `.tabs` div when
we're not drag-and-dropping and to only enable :hover effect when the
class is present.