fix: second part of the fix for correctly moving a tab between group
We were fetching the position of a tab in a group based of its window
index to emit a `TabCreated` request to component of the new group's
tab. But to compute it we were relying on the previous positions of the
tabs, with just the group of the tab changed, not its actual index. The
computation of `getPositionTabInGroup` was in that case incorrect.
Since we already now we want to put it at the end, it's easy to compute.